requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0